% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/with.qibm.R
\docType{methods}
\name{Bias}
\alias{Bias}
\alias{Bias,qibm-method}
\title{Relative Mean Bias}
\usage{
Bias(x, ...)

\S4method{Bias}{qibm}(x, type = c("absolute", "relative", "percent"), log = FALSE, ...)
}
\arguments{
\item{x}{an \R object.}
\item{type}{type of bias to compute.}
\item{log}{whether to compute bias on the original scale for models fit to log-transformed measurements.}
\item{...}{further arguments passed to \code{\link{with}}.}
}
\description{
Relative Mean Bias
}
