% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/with.qibm.R
\docType{methods}
\name{CIndex}
\alias{CIndex}
\alias{CIndex,qibm-method}
\title{Concordance Index}
\usage{
CIndex(x, ...)

\S4method{CIndex}{qibm}(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{...}{further arguments passed to \code{\link{with}}.}
}
\description{
Concordance Index
}
