% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/with.qibm.R
\docType{methods}
\name{Cor}
\alias{Cor}
\alias{Cor,qibm-method}
\title{Between-Method Correlation}
\usage{
Cor(x, ...)

\S4method{Cor}{qibm}(x, diag = FALSE, ...)
}
\arguments{
\item{x}{an \R object.}

\item{...}{further arguments passed to \code{\link{with}}.}

\item{diag}{whether to include correlation matrix diagonals.}
}
\description{
Between-Method Correlation
}
