% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/utils.R
\docType{methods}
\name{describe}
\alias{describe}
\alias{describe,mcmc-method}
\alias{describe,mcmc.list-method}
\title{Descriptive Summaries}
\usage{
describe(x, ...)

\S4method{describe}{mcmc}(x, alpha = 0.05)

\S4method{describe}{mcmc.list}(x, alpha = 0.05)
}
\arguments{
\item{x}{an \R object.}

\item{...}{further arguments passed to or from other methods.}

\item{alpha}{tail probabilities for credible intervals.}
}
\description{
Descriptive Summaries
}
