% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/with.qibm.R
\docType{methods}
\name{ICC}
\alias{ICC}
\alias{ICC,qibm-method}
\title{Intra-Class Correlation Coefficient}
\usage{
ICC(x, ...)

\S4method{ICC}{qibm}(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{...}{further arguments passed to \code{\link{with}}.}
}
\description{
Intra-Class Correlation Coefficient
}
