% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LRMCoef}
\alias{LRMCoef}
\title{Logistic Regression Intercept and Slope}
\usage{
LRMCoef(OR, scale, p, x)
}
\arguments{
\item{OR}{odds ratio for which to derive the intercept and slope.}

\item{scale}{unit increase in \code{x} to which the \code{OR} corresponds.}

\item{p}{event probability at \code{x}.}

\item{x}{value of the predictor variable to which \code{p} corresponds.}
}
\description{
Derive the intercept and slope for a logistic regression model at a specified
odds ratio and event probability.
}
\seealso{
\code{\link{LRM}}.
}
