% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{methods}
\name{print,MCMCDescribe-method}
\alias{print,MCMCDescribe-method}
\alias{print}
\title{Pint Values}
\usage{
\S4method{print}{MCMCDescribe}(x, digits = options()$digits,
  scientific = FALSE)
}
\arguments{
\item{x}{an object returned by \code{\link{describe}}.}

\item{digits}{integer indicated the number of decimal places to display.}

\item{scientific}{whether to encode values in scientific format.}
}
\description{
Pint Values
}
