% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qibm.R
\name{qibm}
\alias{qibm}
\title{Quantitative Imaging Biomarker Model}
\usage{
qibm(fixed, image, operator, data, priors = list(), parameters = c("mu",
  "gamma.img", "Sigma.img", "sigma.opr", "sigma.imgopr", "sigma.err", "gof.rep",
  "gof.obs"), n.burnin = 2500, n.iter = 10000, n.thin = 5, n.chains = 3,
  seed = 123)
}
\arguments{
\item{fixed}{fomula of the form \code{y ~ method} where \code{y} is a
vector of biomarker measurements, or a transformation thereof, and
\code{method} is a vector of identifiers for the imaging methods.}

\item{image}{quoted or unquoted name of a vector of image identifiers.}

\item{operator}{quoted or unquoted name of a vector of operator identifiers.}

\item{data}{data.frame containing the data vectors.}

\item{priors}{list of prior parameter values.  See Details section below for list element names and defaults.}

\item{parameters}{vector naming the model parameters to be returned (default: all).}

\item{n.burnin}{number of MCMC samples to discard as a burn-in sequence.}

\item{n.iter}{total number of samples to generate.}

\item{n.thin}{period at which to save samples.}

\item{n.chains}{number of parallel MCMC chains to generate.}

\item{seed}{numeric random number generator seed.}
}
\value{
A \code{qibm} object that inherits from \code{\link[coda]{mcmc.list}} and
  contains the MCMC sampled model parameter values.
}
\description{
\code{qibm} is used to fit a Bayesian measurement error model for the 
comparison of biomarker measurements derived from different quantitative
imaging methods.
}
\details{
Prior distribution hyperprameters may be specified through the following list elements of the \code{priors} argument.
\describe{
  \item{mu.mean}{Normal mean for mu parameters (default: 0).}
  \item{mu.var}{Normal variance for mu paraemters (defaul: 1e6).}
  \item{sigma2.opr.shape}{Inverse-gamma shape for inter-operator variance parameters (default: 1e-3).}
  \item{sigma2.opr.rate}{Inverse-gamma rate for inter-operator variance parameters (default: 1e-3).}
  \item{sigma.opr.lim}{Two-element vector of uniform minimum and maximum for inter-operator standard deviation parameters (default: inverse-gamma distribution).}
  \item{sigma2.imgopr.shape}{Inverse-gamma shape for image-by-operator variance parameters (default: 1e-3).}
  \item{sigma2.imgopr.rate}{Inverse-gamma rate for image-by-operator variance parameters (default: 1e-3).}
  \item{sigma.imgopr.lim}{Two-element vector of uniform minimum and maximum for image-by-operator standard deviation parameters (default: inverse-gamma distribution).}
  \item{sigma2.err.shape}{Inverse-gamma shape for intra-opearator error variance parameters (default: 1e-3).}
  \item{sigma2.err.rate}{Inverse-gamma rate for intra-opearator error variance parameters (default: 1e-3).}
  \item{sigma.err.lim}{Two-element vector of uniform minimum and maximum for intra-opearator standard deviation  parameters (default: inverse-gamma distribution).}
  \item{Sigma.img.cor}{Scalar pairwise correlation in the correlation matrix defining an inverse-Wishart scale matrix for between-image and method variances (default: 0).}
  \item{Sigma.img.scale}{Scalar multiple of the correlation matrix defining an inverse-Wishart scale matrix for between-image and method variances (default: 1).}
  \item{Sigma.img.df}{Inverse-Wishart degrees of freedom for between-image and method variances (default: number of methods).}
}
}
\examples{
\dontrun{
data(hnc)

## Default prior distributions
fit <- qibm(log(Volume) ~ method, image = lesion, operator = operator,
            data = hnc, n.burnin = 5000, n.iter = 10000, n.thin = 5,
            n.chains = 3)

describe(with(fit, exp(mu)))
describe(with(fit, exp(sqrt(diag(Sigma.img)))))
describe(with(fit, exp(sqrt(sigma.opr^2 + sigma.imgopr^2))))
describe(with(fit, exp(sigma.err)))

describe(Cor(fit))

describe(Bias(fit, log = TRUE))
describe(CIndex(fit))
describe(ICC(fit))

describe(wCV(fit, log = TRUE))
describe(RDC(fit))
describe(RC(fit))

fit.gof <- GOF(fit)
plot(fit.gof)
describe(fit.gof)


## User-specified Uniform(0, 2) priors on standard deviation parameters
fit2 <- qibm(log(Volume) ~ method, image = lesion, operator = operator,
             data = hnc, n.burnin = 5000, n.iter = 10000, n.thin = 5,
             n.chains = 3, priors = list(sigma.opr.lim = c(0, 2),
                                         sigma.imgopr.lim = c(0, 2),
                                         sigma.err.lim = c(0, 2)))
}

}
\seealso{
\code{\link{describe}},
  \code{\link[=with.qibm]{with}},
  \code{\link{Bias}},
  \code{\link{CIndex}},
  \code{\link{Cor}},
  \code{\link{GOF}},
  \code{\link{ICC}},
  \code{\link{LRM}}
  \code{\link{RC}},
  \code{\link{RDC}},
  \code{\link{wCV}}.
}
