% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/with.qibm.R
\docType{methods}
\name{RC}
\alias{RC}
\alias{RC,qibm-method}
\title{Repeatability Coefficient}
\usage{
RC(x, ...)

\S4method{RC}{qibm}(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{...}{further arguments passed to \code{\link{with}}.}
}
\description{
Repeatability Coefficient
}
