% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/with.qibm.R
\docType{methods}
\name{RDC}
\alias{RDC}
\alias{RDC,qibm-method}
\title{Reproducibility Coefficient}
\usage{
RDC(x, ...)

\S4method{RDC}{qibm}(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{...}{further arguments passed to \code{\link{with}}.}
}
\description{
Reproducibility Coefficient
}
