% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.qibm.R
\docType{methods}
\name{transform,qibm-method}
\alias{transform,qibm-method}
\title{Transform Parameters from a Quantitative Imaging Biomarker Model Fit}
\usage{
\S4method{transform}{qibm}(`_data`, f, select, ref = 1, seed = 123, ...)
}
\arguments{
\item{_data}{\code{\link{qibm}} object of MCMC sample model parameter
values.}

\item{f}{transform function to be applied to the model parameters.}

\item{select}{vector of quantification methods to select for the transform
(default: all).}

\item{ref}{identifies the reference quantification method as \code{select[ref]}.}

\item{seed}{numeric random number generator seed.}

\item{...}{further arguments to be passed to \code{f}.}
}
\value{
An \code{\link[coda]{mcmc.list}} object containing transformed MCMC
  sampled model parameter values.
}
\description{
\code{transform} is used to transform model parameters from a \link{qibm} 
fit.
}
