% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R, R/with.qibm.R
\docType{methods}
\name{wCV}
\alias{wCV}
\alias{wCV,qibm-method}
\title{Within-Subject Coefficient of Variation}
\usage{
wCV(x, ...)

\S4method{wCV}{qibm}(x, ...)
}
\arguments{
\item{x}{an \R object.}

\item{...}{further arguments passed to \code{\link{with}}.}
}
\description{
Within-Subject Coefficient of Variation
}
