% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/with.qibm.R
\name{with.qibm}
\alias{with.qibm}
\alias{with}
\title{Evaluate an Expression in Quantitative Imaging Biomarker Model MCMC Samples}
\usage{
\method{with}{qibm}(data, expr, select, ref = 1, ...)
}
\arguments{
\item{data}{\code{\link{qibm}} object of MCMC sample model parameter values.}

\item{expr}{expression to evaluate.}

\item{select}{vector of quantification methods to select for the evaluation
(default: all).}

\item{ref}{identifies the reference quantification method as \code{select[ref]}.}

\item{...}{further arguments to include in the environment.}
}
\value{
An \code{\link[coda]{mcmc.list}} object of MCMC sampled values of the
evaluated \code{expr}.
}
\description{
Evaluate an \R expression in an environment constructed from MCMC
samples, possibly modifying (a copy of) the original samples.
}
\seealso{
\code{\link{qibm}}.
}
