% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qibm.R
\name{qibm}
\alias{qibm}
\title{Quantitative Imaging Biomarker Model}
\usage{
qibm(fixed, image, operator, data, priors = list(), parameters = c("mu",
  "gamma.img", "Sigma.img", "sigma.opr", "sigma.imgopr", "sigma.err", "gof.rep",
  "gof.obs"), n.burnin = 2500, n.iter = 10000, n.thin = 5, n.chains = 3,
  seed = 123)
}
\arguments{
\item{fixed}{fomula of the form \code{y ~ method} where \code{y} is a
vector of biomarker measurements, or a transformation thereof, and
\code{method} is a vector of identifiers for the imaging methods.}

\item{image}{quoted or unquoted name of a vector of image identifiers.}

\item{operator}{quoted or unquoted name of a vector of operator identifiers.}

\item{data}{data.frame containing the data vectors.}

\item{priors}{list of prior parameter values.}

\item{parameters}{vector naming the model parameters to be returned.}

\item{n.burnin}{number of MCMC samples to discard as a burn-in sequence.}

\item{n.iter}{total number of samples to generate.}

\item{n.thin}{period at which to save samples.}

\item{n.chains}{number of parallel MCMC chains to generate.}

\item{seed}{numeric random number generator seed.}
}
\value{
A \code{qibm} object that inherits from \code{\link[coda]{mcmc.list}} and
  contains the MCMC sampled model parameter values.
}
\description{
\code{qibm} is used to fit a Bayesian measurement error model for the 
comparison of biomarker measurements derived from different quantitative
imaging methods.
}
\examples{
\dontrun{
data(hnc)

fit <- qibm(log(Volume) ~ method, image = image, operator = operator,
            data = hnc, n.burnin = 5000, n.iter = 10000, n.thin = 5,
            n.chains = 3)

describe(with(fit, exp(mu)))
describe(with(fit, exp(sqrt(diag(Sigma.img)))))
describe(with(fit, exp(sqrt(sigma.opr^2 + sigma.imgopr^2))))
describe(with(fit, exp(sigma.err)))

describe(Cor(fit))

describe(Bias(fit))
describe(with(fit, exp(mu) - exp(mu[1])))
describe(CIndex(fit))
describe(ICC(fit))

describe(wCV(fit))
describe(RDC(fit))
describe(RC(fit))

fit.gof <- GOF(fit)
plot(fit.gof)
describe(fit.gof)
}

}
\seealso{
\code{\link[=with.qibm]{with}},
  \code{\link{Bias}},
  \code{\link{CIndex}},
  \code{\link{Cor}},
  \code{\link{GOF}},
  \code{\link{ICC}},
  \code{\link{LRM}}
  \code{\link{RC}},
  \code{\link{RDC}},
  \code{\link{wCV}}.
}
