% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qibm.R
\name{qibm}
\alias{qibm}
\title{Quantitative imaging biomarker model}
\usage{
qibm(fixed, image, operator, data, priors = list(), parameters = c("mu",
  "sigma.opr", "sigma.imgopr", "sigma.err"), n.burnin = 2500,
  n.iter = 10000, n.thin = 5, n.chains = 3, seed = 123)
}
\arguments{
\item{fixed}{fomula of the form \code{y ~ method} where \code{y} is a
vector of biomarker measurements, or a transformation thereof, and
\code{method} is a vector of identifiers for the imaging methods.}

\item{image}{quoted or unquoted name of a vector of image identifiers.}

\item{operator}{quoted or unquoted name of a vector of operator identifiers.}

\item{data}{data.frame containing the data vectors.}

\item{priors}{list of prior parameter values.}

\item{parameters}{vector naming the model parameters to be returned.}

\item{n.burnin}{number of MCMC samples to discard as a burn-in sequence.}

\item{n.iter}{total number of samples to generate.}

\item{n.thin}{period at which to save samples.}

\item{n.chains}{number of parallel MCMC chains to generate.}

\item{seed}{numeric random number generator seed.}
}
\value{
An \code{\link[coda]{mcmc.list}} object containing the MCMC sampled
  model parameter values.
}
\description{
\code{qibm} is used to fit a Bayesian measurement error model for the 
comparison of biomarker measurements derived from different quantitative
imaging methods.
}
